change the quantity type from integer to float

+++ database/migrations/2024_11_22_082239_create_store_order_items_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('store_order_id')->constrained('store_orders')->cascadeOnDelete();
             $table->foreignId('product_inventory_id')->constrained('product_inventories')->cascadeOnDelete();
-            $table->integer('quantity_ordered');
-            $table->integer('quantity_approved')->default(0);
-            $table->integer('quantity_received')->default(0);
+            $table->float('quantity_ordered');
+            $table->float('quantity_approved')->default(0);
+            $table->float('quantity_received')->default(0);
             $table->double('total_cost');
             $table->text('remarks')->nullable();
             $table->timestamps();
